[WebCodecs] Improve errors when operating on closed VideoFrame.
This adds exceptions when texturing or drawing a closed VideoFrame
to canvas via WebGL or 2D canvas respectively. These are clear
programming errors that should be called out and not operational
errors that might occur in the wild.

It's unclear if this is the problem of the linked issue, but it
will certainly help users track down issues if so.

Bug: 1236821
Change-Id: I590aa020b74483b58602e5237582889bc5e3cd3e